requirements.txt
# Core dependencies
chromadb>=0.4.0
pyyaml>=6.0
pillow>=10.0.0

# Optional but recommended
mutagen>=1.46.0

# For LLM integration
ollama>=0.1.0

# Logging and utilities
python-dotenv>=1.0.0
